FIX#5 merge bnechmarks into one file
will be usefull when we have a datacontainer
